refactor(graph): move the drill-down members to GroupingMixin
GroupingMixin already owns enterGroup and exitGroup, the two entry points of drill-down navigation, and it already had
to declare isValidRoot and getCurrentRoot as external dependencies to implement them. Move the members those methods
depend on: resetViewOnRootChange, getCurrentRoot, getTranslateForRoot, getChildOffsetForCell, home and isValidRoot.

AbstractGraph keeps calling home and resetViewOnRootChange from processChange, and getCurrentRoot from
getDefaultParent, which declaration merging supports without any change at the call sites.

defaultParent, getDefaultParent and setDefaultParent stay deliberately. They are the most used entry point of the whole
API and are read from processChange, so the cohesion gain does not justify the churn.

Worth recording for the eventual conversion of this mixin to a plugin: drill-down is view navigation while grouping is
model restructuring. They are two concerns sharing a file because no better home exists today, so this move should not
be read as a statement that they belong together.

AbstractGraph loses 89 lines and its last use of Point, and now stands at 811 lines, down from 1337 at the start of the
plan.
